RUM-3433 refactor plugin to make it testable
