+++ integrations/ktor/src/commonMain/kotlin/com/datadog/kmp/ktor/DatadogKtorPlugin.kt
@@ -6,27 +6,15 @@
 
 package com.datadog.kmp.ktor
 
-import com.benasher44.uuid.uuid4
+import com.datadog.kmp.ktor.internal.plugin.DatadogKtorPlugin
+import com.datadog.kmp.ktor.internal.plugin.buildClientPlugin
 import com.datadog.kmp.ktor.internal.trace.DefaultSpanIdGenerator
 import com.datadog.kmp.ktor.internal.trace.DefaultTraceIdGenerator
-import com.datadog.kmp.ktor.internal.trace.SpanIdGenerator
-import com.datadog.kmp.ktor.internal.trace.TraceIdGenerator
 import com.datadog.kmp.rum.RumMonitor
-import com.datadog.kmp.rum.RumResourceKind
-import com.datadog.kmp.rum.RumResourceMethod
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.api.ClientPlugin
-import io.ktor.client.plugins.api.createClientPlugin
-import io.ktor.client.statement.request
-import io.ktor.http.HttpMethod
-import io.ktor.util.AttributeKey
 
-internal const val PLUGIN_NAME = "Datadog"
-internal const val DD_REQUEST_ID = "X-Datadog-Request-ID"
-internal val DD_REQUEST_ID_ATTR = AttributeKey<String>(DD_REQUEST_ID)
 internal const val DEFAULT_TRACE_SAMPLE_RATE: Float = 20f
-internal const val MIN_SAMPLE_RATE: Double = 0.0
-internal const val MAX_SAMPLE_RATE: Double = 100.0
 
 /**
  * Create a Datadog plugin for a Ktor [HttpClient].
@@ -38,85 +26,12 @@ internal const val MAX_SAMPLE_RATE: Double = 100.0
 fun datadogKtorPlugin(
     tracedHosts: Map<String, Set<TracingHeaderType>> = emptyMap(),
     traceSamplingRate: Float = DEFAULT_TRACE_SAMPLE_RATE
-): ClientPlugin<Unit> = datadogKtorPlugin(
-    tracedHosts,
-    traceSamplingRate,
-    DefaultTraceIdGenerator(),
-    DefaultSpanIdGenerator()
-)
-
-internal fun datadogKtorPlugin(
-    tracedHosts: Map<String, Set<TracingHeaderType>>,
-    traceSamplingRate: Float,
-    traceIdGenerator: TraceIdGenerator,
-    spanIdGenerator: SpanIdGenerator
 ): ClientPlugin<Unit> {
-    return createClientPlugin(PLUGIN_NAME) {
-        // TODO RUM-5228 report request timings (DNS, SSL, …)
-        // TODO RUM-5229 report request exceptions
-
-        onRequest { request, _ ->
-
-            val isSampledIn = RNG.nextDouble(MIN_SAMPLE_RATE, MAX_SAMPLE_RATE).toFloat() < traceSamplingRate
-            val traceHeaderTypes = tracedHosts[request.url.host]
-            val attributes = mutableMapOf<String, Any?>()
-
-            val traceId = traceIdGenerator.generateTraceId()
-            val spanId = spanIdGenerator.generateSpanId()
-
-            if (isSampledIn && !traceHeaderTypes.isNullOrEmpty()) {
-                traceHeaderTypes.forEach { headerType ->
-                    headerType.injectHeaders(request, true, traceId, spanId)
-                }
-                attributes[RUM_TRACE_ID] = traceId.toHexString()
-                attributes[RUM_SPAN_ID] = spanId.raw.toString()
-                attributes[RUM_RULE_PSR] = traceSamplingRate
-            } else {
-                TracingHeaderType.entries.forEach { headerType ->
-                    headerType.injectHeaders(request, false, traceId, spanId)
-                }
-            }
-
-            val requestId = uuid4().toString()
-            request.attributes.put(DD_REQUEST_ID_ATTR, requestId)
-            RumMonitor.get().startResource(
-                key = requestId,
-                method = request.method.asRumMethod(),
-                url = request.url.buildString(),
-                attributes = emptyMap()
-            )
-        }
-
-        onResponse { response ->
-            val requestId = response.request.attributes.getOrNull(DD_REQUEST_ID_ATTR)
-            if (requestId != null) {
-                RumMonitor.get().stopResource(
-                    key = requestId,
-                    statusCode = response.status.value,
-                    size = null, // TODO RUM-5233 report request size
-                    kind = RumResourceKind.NATIVE,
-                    attributes = emptyMap()
-                )
-            } else {
-                // TODO RUM-5254 handle missing request id case
-            }
-        }
-    }
-}
-
-private fun HttpMethod.asRumMethod(): RumResourceMethod {
-    return when (this) {
-        HttpMethod.Post -> RumResourceMethod.POST
-        HttpMethod.Get -> RumResourceMethod.GET
-        HttpMethod.Head -> RumResourceMethod.HEAD
-        HttpMethod.Put -> RumResourceMethod.PUT
-        HttpMethod.Delete -> RumResourceMethod.DELETE
-        HttpMethod.Patch -> RumResourceMethod.PATCH
-        HttpMethod.Options -> RumResourceMethod.OPTIONS
-
-        else -> {
-            // TODO log unknown HTTP method
-            RumResourceMethod.CONNECT
-        }
-    }
+    return DatadogKtorPlugin(
+        RumMonitor.get(),
+        tracedHosts,
+        traceSamplingRate,
+        DefaultTraceIdGenerator(),
+        DefaultSpanIdGenerator()
+    ).buildClientPlugin()
 }
